requirements.txt
dash>=2.0.0
plotly>=5.4.0
dash-daq>=0.5.0
matplotlib>=3.3.3
termcolor>=1.1.0
numpy>=1.18.4
lmdb>=1.1.1
lz4>=3.1.3
msgpack>=1.0.2
msgpack_numpy>=0.4.7.1
faiss-cpu>=1.7.1
scikit-learn>=0.23.2
tqdm
